merging to have exception handling code

+++ quora-service/src/main/java/com/upgrad/quora/service/dao/UserDao.java
@@ -1,11 +1,15 @@
 package com.upgrad.quora.service.dao;
 
+import com.upgrad.quora.service.entity.UserAuthTokenEntity;
 import com.upgrad.quora.service.entity.UserEntity;
+import com.upgrad.quora.service.exception.UserNotFoundException;
 
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Repository;
 
-
 import javax.persistence.EntityManager;
+import javax.persistence.NoResultException;
 import javax.persistence.PersistenceContext;
 
 @Repository
@@ -14,17 +18,44 @@ public class UserDao {
     @PersistenceContext
     private EntityManager entityManager;
     
-    public void deleteUser(final String userUuid) {
+    public String deleteUser(final String userUuid) throws UserNotFoundException {
 
-        UserEntity userEntity = entityManager.createNamedQuery("userByUuid", UserEntity.class).setParameter("uuid", userUuid).getSingleResult();
-        entityManager.remove(userEntity);
+        UserEntity userEntity;
+        try {
+            userEntity = entityManager.createNamedQuery("userByUuid", UserEntity.class).setParameter("uuid", userUuid).getSingleResult();
+        }
+        catch(NoResultException nre) {
+            userEntity = null;
+        }
+
+        if(userEntity == null) {
+            throw new UserNotFoundException("USR-001", "User with entered uuid to be deleted does not exist");
+        }
+        else {
+            entityManager.remove(userEntity);
+            return userUuid;
+        }
+
+
+
 
     }
 
     public UserEntity viewUserProfile(final String userUuid) {
 
-        return entityManager.createNamedQuery("userByUuid", UserEntity.class).setParameter("uuid", userUuid).getSingleResult();
+        try {
+            return entityManager.createNamedQuery("userByUuid", UserEntity.class).setParameter("uuid", userUuid).getSingleResult();
+        } catch(NoResultException nre) {
+            return null;
+        }
 
+    }
 
+    public UserAuthTokenEntity getUserAuthToken(final String accesstoken) {
+        try {
+            return entityManager.createNamedQuery("userAuthTokenByAccessToken", UserAuthTokenEntity.class).setParameter("accessToken", accesstoken).getSingleResult();
+        } catch (NoResultException nre) {
+            return null;
+        }
     }
 }
